MIK-224 replaced doubles of prices into decimal

+++ MikietaApi.Tests/Services/DeliveryServiceTest.cs
@@ -5,16 +5,16 @@ namespace MikietaApi.Tests.Services;
 
 public class DeliveryServiceTest
 {
-    [TestCase(1d, 2d, ExpectedResult = 2d)]
-    [TestCase(1.1d, 2d, ExpectedResult = 2d)]
-    [TestCase(1.99d, 2d, ExpectedResult = 2d)]
-    [TestCase(0.9d, 2d, ExpectedResult = 0d)]
-    [TestCase(2.9d, 4d, ExpectedResult = 8d)]
-    [TestCase(2.9d, 0d, ExpectedResult = 0d)]
-    public double CalculateDeliveryPrice(double distance, double deliveryPrice)
+    [TestCase(1d, 2, ExpectedResult = 2)]
+    [TestCase(1.1d, 2, ExpectedResult = 2)]
+    [TestCase(1.99d, 2, ExpectedResult = 2)]
+    [TestCase(0.9d, 2, ExpectedResult = 0)]
+    [TestCase(2.9d, 4, ExpectedResult = 8)]
+    [TestCase(2.9d, 0, ExpectedResult = 0)]
+    public decimal CalculateDeliveryPrice(double distance, decimal deliveryPrice)
     {
         var service = new DeliveryService(new DataContextMock(), new GoogleLocationService());
-        
-        return Helpers.InvokePrivateMethod<double>(service, "CalculateDeliveryPrice", distance, deliveryPrice);;
+
+        return Helpers.InvokePrivateMethod<decimal>(service, "CalculateDeliveryPrice", distance, deliveryPrice);;
     }
 }
