+++ MikietaApi.Tests/Stripe/StripeRequestConverterTest.cs
@@ -8,7 +8,7 @@ namespace MikietaApi.Tests.Stripe;
 public class StripeRequestConverterTest
 {
     private static OrderedProductOrderedIngredientEntity CreateOrderedProductOrderedIngredient(int quantity,
-        double small, double medium, double large)
+        decimal small, decimal medium, decimal large)
     {
         return new OrderedProductOrderedIngredientEntity
         {
@@ -24,33 +24,33 @@ public class StripeRequestConverterTest
 
     private static IEnumerable<TestCaseData> Convert_Cases()
     {
-        yield return new TestCaseData(12d, 1, PizzaType.Small,
+        yield return new TestCaseData(12m, 1, PizzaType.Small,
                 new[] { CreateOrderedProductOrderedIngredient(1, 1, 2, 3) }, 1300)
             .SetName("Convert 01");
 
-        yield return new TestCaseData(12d, 2, PizzaType.Small,
+        yield return new TestCaseData(12m, 2, PizzaType.Small,
                 new[] { CreateOrderedProductOrderedIngredient(1, 1, 2, 3) }, 1300)
             .SetName("Convert 02");
 
-        yield return new TestCaseData(12.54d, 1, PizzaType.Medium,
+        yield return new TestCaseData(12.54m, 1, PizzaType.Medium,
                 new[] { CreateOrderedProductOrderedIngredient(1, 1, 2, 3) }, 1454)
             .SetName("Convert 03");
 
-        yield return new TestCaseData(1d, 1, PizzaType.Large,
+        yield return new TestCaseData(1m, 1, PizzaType.Large,
                 new[] { CreateOrderedProductOrderedIngredient(1, 1, 2, 3) }, 400)
             .SetName("Convert 04");
 
-        yield return new TestCaseData(1d, 1, null,
+        yield return new TestCaseData(1m, 1, null,
                 new[] { CreateOrderedProductOrderedIngredient(1, 1, 2, 3) }, 100)
             .SetName("Convert 05");
         
-        yield return new TestCaseData(1d, 1, PizzaType.Small,
+        yield return new TestCaseData(1m, 1, PizzaType.Small,
                 new[] { CreateOrderedProductOrderedIngredient(2, 1, 2, 3) }, 300)
             .SetName("Convert 06");
     }
 
     [TestCaseSource(nameof(Convert_Cases))]
-    public void Convert(double price, int quantity, PizzaType? pizzaType,
+    public void Convert(decimal price, int quantity, PizzaType? pizzaType,
         OrderedProductOrderedIngredientEntity[] orderedProductOrderedIngredients,
         int expectedPrice)
     {
