+++ MikietaApi/Data/Entities/ProductEntity.cs
@@ -11,7 +11,7 @@ public class ProductEntity
     public string Name { get; set; } = null!;
     public ICollection<IngredientEntity> Ingredients { get; set; }
     public ICollection<PizzaSizeEntity> Sizes { get; set; }
-    public double? Price { get; set; }
+    public decimal? Price { get; set; }
     public ProductType ProductType { get; set; }
     public string? Description { get; set; }
     public bool IsDeleted { get; set; }
@@ -20,7 +20,7 @@ public class ProductEntity
     public ImageEntity? Image { get; set; }
     public int Index { get; set; }
 
-    public double? GetPrice(PizzaType? type)
+    public decimal? GetPrice(PizzaType? type)
     {
         return type.HasValue ? Sizes.First(x => x.Size == type).Price : Price;
     }
